+++ test/fuzzy-search.test.js
@@ -19,7 +19,6 @@ const defaultOptions = {
   shouldSort: true,
   getFn: get,
   sortFn: (a, b) => (a.score - b.score),
-  tokenize: false,
   matchAllTokens: false,
   includeMatches: false,
   includeScore: false,
@@ -28,7 +27,7 @@ const defaultOptions = {
 
 const setup = (itemList, overwriteOptions) => {
   const list = itemList || defaultList
-  const options = {...defaultOptions, ...overwriteOptions}
+  const options = { ...defaultOptions, ...overwriteOptions }
 
   return new Fuse(list, options)
 }
@@ -38,7 +37,7 @@ describe('Flat list of strings: ["Apple", "Orange", "Banana"]', () => {
   beforeEach(() => fuse = setup())
 
   it('should have the correct configuration', () => {
-    const expected = {list: defaultList, options: defaultOptions}
+    const expected = { list: defaultList, options: defaultOptions }
     expect(fuse).toMatchObject(expected)
   })
 
@@ -89,7 +88,7 @@ describe('Flat list of strings: ["Apple", "Orange", "Banana"]', () => {
 
   describe('When performing a fuzzy search for the term "nan" with a limit of 1 result', () => {
     let result
-    beforeEach(() => result = fuse.search('nan', {limit: 1}))
+    beforeEach(() => result = fuse.search('nan', { limit: 1 }))
 
     test('we get a list of containing 1 item: [2]', () => {
       expect(result).toHaveLength(1)
@@ -101,102 +100,19 @@ describe('Flat list of strings: ["Apple", "Orange", "Banana"]', () => {
   })
 })
 
-describe('List of books - searching "title" and "author"', () => {
-  let fuse
-  beforeEach(() => fuse = setup(books, {keys: ['title', 'author'], tokenize: true}))
-
-  it('should have the correct configuration', () => {
-    const expected = {list: books, options: {...defaultOptions, keys: ['title', 'author'], tokenize: true}}
-    expect(fuse).toMatchObject(expected)
-  })
-
-  describe('When searching for the term "HTML5"', () => {
-    let result
-    beforeEach(() => result = fuse.search('HTML5'))
-
-    test('we get a list of containing 3 items', () => {
-      expect(result).toHaveLength(3)
-    })
-
-    test('and the first item has the matching key/value pairs', () => {
-      expect(result[0]).toMatchObject({title: 'HTML5', author: 'Remy Sharp'})
-    })
-  })
-
-  describe('When searching with pattern length over 32', () => {
-    let result
-    beforeEach(() => result = fuse.search(Array(16).fill('HTML5').join(' ')))
-
-    test('we get a list of containing 3 items', () => {
-      expect(result).toHaveLength(3)
-    })
-
-    test('and the first item has the matching key/value pairs', () => {
-      expect(result[0]).toMatchObject({title: 'HTML5', author: 'Remy Sharp'})
-    })
-  })
-
-  describe('When searching for the term "Jeeves Woodhouse"', () => {
-    let result
-    beforeEach(() => result = fuse.search('Jeeves Woodhouse'))
-
-    test('we get a list of containing 6 items', () => {
-      expect(result).toHaveLength(6)
-    })
-
-    test('which are all the books written by "P.D. Woodhouse"', () => {
-      const expectedResult = [
-        {title: 'Right Ho Jeeves', author: 'P.D. Woodhouse'},
-        {title: 'Thank You Jeeves', author: 'P.D. Woodhouse'},
-        {title: 'The Code of the Wooster', author: 'P.D. Woodhouse'},
-        {title: 'The Lock Artist', author: 'Steve Hamilton'},
-        {title: 'the wooster code', author: 'aa'},
-        {title: 'The code of the wooster', author: 'aa'}
-      ]
-      expect(result).toStrictEqual(expectedResult)
-    })
-  })
-
-  describe('When searching for the term "brwn"', () => {
-    let result
-    beforeEach(() => result = fuse.search('brwn'))
-
-    test('we get a list containing at least 3 items', () => {
-      expect(result.length).toBeGreaterThanOrEqual(3)
-    })
-
-    test('and the first 3 items should be all the books written by Dan Brown"', () => {
-      expect(result[0]).toMatchObject({
-        'title': 'The DaVinci Code',
-        'author': 'Dan Brown'
-      })
-
-      expect(result[1]).toMatchObject({
-        'title': 'Angels & Demons',
-        'author': 'Dan Brown'
-      })
-
-      expect(result[2]).toMatchObject({
-        'title': 'The Lost Symbol',
-        'author': 'Dan Brown'
-      })
-    })
-  })
-})
-
 describe('Deep key search, with ["title", "author.firstName"]', () => {
   const customBookList = [{
     title: "Old Man's War",
-    author: {firstName: 'John', lastName: 'Scalzi'}
+    author: { firstName: 'John', lastName: 'Scalzi' }
   }, {
     title: 'The Lock Artist',
-    author: {firstName: 'Steve', lastName: 'Hamilton'}
-  }, {title: 'HTML5'}, {title: 'A History of England', author: {firstName: 1066, lastName: 'Hastings'}}]
+    author: { firstName: 'Steve', lastName: 'Hamilton' }
+  }, { title: 'HTML5' }, { title: 'A History of England', author: { firstName: 1066, lastName: 'Hastings' } }]
   let fuse
-  beforeEach(() => fuse = setup(customBookList, {keys: ['title', 'author.firstName']}))
+  beforeEach(() => fuse = setup(customBookList, { keys: ['title', 'author.firstName'] }))
 
   it('should have the correct configuration', () => {
-    const expected = {list: customBookList, options: {...defaultOptions, keys: ['title', 'author.firstName']}}
+    const expected = { list: customBookList, options: { ...defaultOptions, keys: ['title', 'author.firstName'] } }
     expect(fuse).toMatchObject(expected)
   })
 
@@ -211,7 +127,7 @@ describe('Deep key search, with ["title", "author.firstName"]', () => {
     test('and the first item has the matching key/value pairs', () => {
       expect(result[0]).toMatchObject({
         title: 'The Lock Artist',
-        author: {firstName: 'Steve', lastName: 'Hamilton'}
+        author: { firstName: 'Steve', lastName: 'Hamilton' }
       })
     })
   })
@@ -227,7 +143,7 @@ describe('Deep key search, with ["title", "author.firstName"]', () => {
     test('whose value matches', () => {
       expect(result[0]).toMatchObject({
         title: 'A History of England',
-        author: {firstName: 1066, lastName: 'Hastings'}
+        author: { firstName: 1066, lastName: 'Hastings' }
       })
     })
   })
@@ -259,7 +175,7 @@ describe('Custom search function, with ["title", "author.firstName"]', () => {
   beforeEach(() => fuse = setup(customBookList, customOptions))
 
   it('should have the correct configuration', () => {
-    const expected = {list: customBookList, options: {...defaultOptions, ...customOptions}}
+    const expected = { list: customBookList, options: { ...defaultOptions, ...customOptions } }
     expect(fuse).toMatchObject(expected)
   })
 
@@ -274,7 +190,7 @@ describe('Custom search function, with ["title", "author.firstName"]', () => {
     test('and the first item has the matching key/value pairs', () => {
       expect(result[0]).toMatchObject({
         title: 'The Lock Artist',
-        author: {firstName: 'Steve', lastName: 'Hamilton'}
+        author: { firstName: 'Steve', lastName: 'Hamilton' }
       })
     })
   })
@@ -291,10 +207,10 @@ describe('Custom search function, with ["title", "author.firstName"]', () => {
 
 describe('Include score in result list: ["Apple", "Orange", "Banana"]', () => {
   let fuse
-  beforeEach(() => fuse = setup(defaultList, {includeScore: true}))
+  beforeEach(() => fuse = setup(defaultList, { includeScore: true }))
 
   it('should have the correct configuration', () => {
-    const expected = {list: defaultList, options: {...defaultOptions, includeScore: true}}
+    const expected = { list: defaultList, options: { ...defaultOptions, includeScore: true } }
     expect(fuse).toMatchObject(expected)
   })
 
@@ -347,7 +263,7 @@ describe('Only include ID in result list, with "ISBN"', () => {
   beforeEach(() => fuse = setup(customBookList, customOptions))
 
   it('should have the correct configuration', () => {
-    const expected = {list: customBookList, options: {...defaultOptions, ...customOptions}}
+    const expected = { list: customBookList, options: { ...defaultOptions, ...customOptions } }
     expect(fuse).toMatchObject(expected)
   })
 
@@ -384,7 +300,7 @@ describe('Include both ID and score in results list', () => {
   beforeEach(() => fuse = setup(customBookList, customOptions))
 
   it('should have the correct configuration', () => {
-    const expected = {list: customBookList, options: {...defaultOptions, ...customOptions}}
+    const expected = { list: customBookList, options: { ...defaultOptions, ...customOptions } }
     expect(fuse).toMatchObject(expected)
   })
 
@@ -425,7 +341,7 @@ describe('Search when IDs are numbers', () => {
   beforeEach(() => fuse = setup(customBookList, customOptions))
 
   it('should have the correct configuration', () => {
-    const expected = {list: customBookList, options: {...defaultOptions, ...customOptions}}
+    const expected = { list: customBookList, options: { ...defaultOptions, ...customOptions } }
     expect(fuse).toMatchObject(expected)
   })
 
@@ -474,7 +390,7 @@ describe('Recurse into arrays', () => {
   beforeEach(() => fuse = setup(customBookList, customOptions))
 
   it('should have the correct configuration', () => {
-    const expected = {list: customBookList, options: {...defaultOptions, ...customOptions}}
+    const expected = { list: customBookList, options: { ...defaultOptions, ...customOptions } }
     expect(fuse).toMatchObject(expected)
   })
 
@@ -491,7 +407,7 @@ describe('Recurse into arrays', () => {
     })
 
     test('with matched tag provided', () => {
-      const {matches} = result[0]
+      const { matches } = result[0]
       expect(matches[0]).toMatchObject({
         key: 'tags',
         arrayIndex: 1,
@@ -540,7 +456,7 @@ describe('Recurse into objects in arrays', () => {
   beforeEach(() => fuse = setup(customBookList, customOptions))
 
   it('should have the correct configuration', () => {
-    const expected = {list: customBookList, options: {...defaultOptions, ...customOptions}}
+    const expected = { list: customBookList, options: { ...defaultOptions, ...customOptions } }
     expect(fuse).toMatchObject(expected)
   })
 
@@ -576,7 +492,7 @@ describe('Searching by ID', () => {
   beforeEach(() => fuse = setup(customBookList, customOptions))
 
   it('should have the correct configuration', () => {
-    const expected = {list: customBookList, options: {...defaultOptions, ...customOptions}}
+    const expected = { list: customBookList, options: { ...defaultOptions, ...customOptions } }
     expect(fuse).toMatchObject(expected)
   })
 
@@ -597,11 +513,11 @@ describe('Searching by ID', () => {
 
 describe('Searching by nested ID', () => {
   const customBookList = [{
-    'ISBN': {'name': 'A'},
+    'ISBN': { 'name': 'A' },
     'title': "Old Man's War",
     'author': 'John Scalzi'
   }, {
-    'ISBN': {'name': 'B'},
+    'ISBN': { 'name': 'B' },
     'title': 'The Lock Artist',
     'author': 'Steve Hamilton'
   }]
@@ -613,7 +529,7 @@ describe('Searching by nested ID', () => {
   beforeEach(() => fuse = setup(customBookList, customOptions))
 
   it('should have the correct configuration', () => {
-    const expected = {list: customBookList, options: {...defaultOptions, ...customOptions}}
+    const expected = { list: customBookList, options: { ...defaultOptions, ...customOptions } }
     expect(fuse).toMatchObject(expected)
   })
 
@@ -642,7 +558,7 @@ describe('Set new list on Fuse', () => {
   })
 
   it('should have the correct configuration', () => {
-    const expected = {list: vegetables, options: defaultOptions}
+    const expected = { list: vegetables, options: defaultOptions }
     expect(fuse).toMatchObject(expected)
   })
 
@@ -687,8 +603,8 @@ describe('Weighted search', () => {
   describe('When searching for the term "John Smith" with author weighted higher', () => {
     const customOptions = {
       keys: [
-        {name: 'title', weight: 0.3},
-        {name: 'author', weight: 0.7}
+        { name: 'title', weight: 0.3 },
+        { name: 'author', weight: 0.7 }
       ]
     }
     let fuse
@@ -699,15 +615,15 @@ describe('Weighted search', () => {
     })
 
     test('We get the the exactly matching object', () => {
-      expect(result[0]).toMatchObject({title: 'The life of Jane', author: 'John Smith', tags: ['john', 'smith']})
+      expect(result[0]).toMatchObject({ title: 'The life of Jane', author: 'John Smith', tags: ['john', 'smith'] })
     })
   })
 
   describe('When searching for the term "John Smith" with title weighted higher', () => {
     const customOptions = {
       keys: [
-        {name: 'title', weight: 0.7},
-        {name: 'author', weight: 0.3}
+        { name: 'title', weight: 0.7 },
+        { name: 'author', weight: 0.3 }
       ]
     }
     let fuse
@@ -718,15 +634,15 @@ describe('Weighted search', () => {
     })
 
     test('We get the the exactly matching object', () => {
-      expect(result[0]).toMatchObject({title: 'John Smith', author: 'Steve Pearson', tags: ['steve', 'pearson']})
+      expect(result[0]).toMatchObject({ title: 'John Smith', author: 'Steve Pearson', tags: ['steve', 'pearson'] })
     })
   })
 
   describe('When searching for the term "Man", where the author is weighted higher than title', () => {
     const customOptions = {
       keys: [
-        {name: 'title', weight: 0.3},
-        {name: 'author', weight: 0.7}
+        { name: 'title', weight: 0.3 },
+        { name: 'author', weight: 0.7 }
       ]
     }
     let fuse
@@ -737,15 +653,15 @@ describe('Weighted search', () => {
     })
 
     test('We get the the exactly matching object', () => {
-      expect(result[0]).toMatchObject({title: 'Right Ho Jeeves', author: 'P.D. Mans', tags: ['fiction', 'war']})
+      expect(result[0]).toMatchObject({ title: 'Right Ho Jeeves', author: 'P.D. Mans', tags: ['fiction', 'war'] })
     })
   })
 
   describe('When searching for the term "Man", where the title is weighted higher than author', () => {
     const customOptions = {
       keys: [
-        {name: 'title', weight: 0.7},
-        {name: 'author', weight: 0.3}
+        { name: 'title', weight: 0.7 },
+        { name: 'author', weight: 0.3 }
       ]
     }
     let fuse
@@ -756,16 +672,16 @@ describe('Weighted search', () => {
     })
 
     test('We get the the exactly matching object', () => {
-      expect(result[0]).toMatchObject({title: "Old Man's War fiction", author: 'John X', tags: ['war']})
+      expect(result[0]).toMatchObject({ title: "Old Man's War fiction", author: 'John X', tags: ['war'] })
     })
   })
 
   describe('When searching for the term "War", where tags are weighted higher than all other keys', () => {
     const customOptions = {
       keys: [
-        {name: 'title', weight: 0.4},
-        {name: 'author', weight: 0.1},
-        {name: 'tags', weight: 0.5}
+        { name: 'title', weight: 0.4 },
+        { name: 'author', weight: 0.1 },
+        { name: 'tags', weight: 0.5 }
       ]
     }
     let fuse
@@ -776,13 +692,13 @@ describe('Weighted search', () => {
     })
 
     test('We get the the exactly matching object', () => {
-      expect(result[0]).toMatchObject({title: "Old Man's War fiction", author: 'John X', tags: ['war']})
+      expect(result[0]).toMatchObject({ title: "Old Man's War fiction", author: 'John X', tags: ['war'] })
     })
   })
 })
 
 describe('Search location', () => {
-  const customList = [{name: 'Hello World'}]
+  const customList = [{ name: 'Hello World' }]
   const customOptions = {
     keys: ['name'],
     includeScore: true,
@@ -811,54 +727,11 @@ describe('Search location', () => {
   })
 })
 
-describe('Search with match all tokens', () => {
-  const customList = [
-    'AustralianSuper - Corporate Division',
-    'Aon Master Trust - Corporate Super',
-    'Promina Corporate Superannuation Fund',
-    'Workforce Superannuation Corporate',
-    'IGT (Australia) Pty Ltd Superannuation Fund'
-  ]
-  let fuse
-
-  beforeEach(() => fuse = setup(customList, {tokenize: true}))
-
-  describe('When searching for the term "Australia"', () => {
-    let result
-    beforeEach(() => result = fuse.search('Australia'))
-
-    test('We get a list containing exactly 2 items', () => {
-      expect(result).toHaveLength(2)
-    })
-
-    test('whose items represent the indices of "AustralianSuper - Corporate Division" and "IGT (Australia) Pty Ltd Superannuation Fund"', () => {
-      expect(result).toContain(0)
-      expect(result).toContain(4)
-    })
-  })
-
-  describe('When searching for the term "corporate"', () => {
-    let result
-    beforeEach(() => result = fuse.search('corporate'))
-
-    test('We get a list containing exactly 4 items', () => {
-      expect(result).toHaveLength(4)
-    })
-
-    test('whose items represent the indices of "AustralianSuper - Corporate Division", "Aon Master Trust - Corporate Super", "Promina Corporate Superannuation Fund" and "Workforce Superannuation Corporate"', () => {
-      expect(result).toContain(0)
-      expect(result).toContain(1)
-      expect(result).toContain(2)
-      expect(result).toContain(3)
-    })
-  })
-})
-
 describe('Searching with default options', () => {
   const customList = ['t te tes test tes te t']
   let fuse
 
-  beforeEach(() => fuse = new Fuse(customList, {includeMatches: true}))
+  beforeEach(() => fuse = new Fuse(customList, { includeMatches: true }))
 
   describe('When searching for the term "test"', () => {
     let result
@@ -886,7 +759,7 @@ describe('Searching with findAllMatches', () => {
   const customList = ['t te tes test tes te t']
   let fuse
 
-  beforeEach(() => fuse = new Fuse(customList, {includeMatches: true, findAllMatches: true}))
+  beforeEach(() => fuse = new Fuse(customList, { includeMatches: true, findAllMatches: true }))
 
   describe('When searching for the term "test"', () => {
     let result
@@ -907,7 +780,7 @@ describe('Searching with minCharLength', () => {
   const customList = ['t te tes test tes te t']
   let fuse
 
-  beforeEach(() => fuse = new Fuse(customList, {includeMatches: true, minMatchCharLength: 2}))
+  beforeEach(() => fuse = new Fuse(customList, { includeMatches: true, minMatchCharLength: 2 }))
 
   describe('When searching for the term "test"', () => {
     let result
@@ -938,15 +811,15 @@ describe('Sorted search results', () => {
   const customList = [
     {
       title: 'Right Ho Jeeves',
-      author: {firstName: 'P.D', lastName: 'Woodhouse'}
+      author: { firstName: 'P.D', lastName: 'Woodhouse' }
     },
     {
       title: 'The Code of the Wooster',
-      author: {firstName: 'P.D', lastName: 'Woodhouse'}
+      author: { firstName: 'P.D', lastName: 'Woodhouse' }
     },
     {
       title: 'Thank You Jeeves',
-      author: {firstName: 'P.D', lastName: 'Woodhouse'}
+      author: { firstName: 'P.D', lastName: 'Woodhouse' }
     }
   ]
   const customOptions = {
@@ -1004,7 +877,6 @@ describe('Searching using string large strings', () => {
     text: 'super+large+much+unique+36+very+wow+'
   }]
   const options = {
-    include: ['score', 'matches'],
     shouldSort: true,
     threshold: 0.5,
     location: 0,