+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# Version 4.0.0
+
+- Added extended search:
+
+| Token       | Match type                 | Description                            |
+| ----------- | -------------------------- | -------------------------------------- |
+| `jscript`   | fuzzy-match                | Items that match `jscript`             |
+| `'python`   | exact-match                | Items that include `python`            |
+| `!ruby`     | inverse-exact-match        | Items that do not include `ruby`       |
+| `^java`     | prefix-exact-match         | Items that start with `java`           |
+| `!^earlang` | inverse-prefix-exact-match | Items that do not start with `earlang` |
+| `.js$`      | suffix-exact-match         | Items that end with `.js`              |
+| `!.go$`     | inverse-suffix-exact-match | Items that do not end with `.go`       |
+
+- Removed tokenization
+
 # Version 3.6.0
 
 - Improved error handling for keys